Add optics & vehicles compats to realistic names
