+++ addons/realisticnames/compat_optics/CfgWeapons.hpp
@@ -1,0 +1,33 @@
+class CfgWeapons {
+    class optic_Hamr;
+    class ACE_optic_Hamr_2D: optic_Hamr {
+        displayName = CSTRING(optic_hamr_2d);
+    };
+    class ACE_optic_Hamr_PIP: ACE_optic_Hamr_2D {
+        displayName = CSTRING(optic_hamr_pip);
+    };
+
+    class optic_Arco;
+    class ACE_optic_Arco_2D: optic_Arco {
+        displayName = CSTRING(optic_arco_2d);
+    };
+    class ACE_optic_Arco_PIP: ACE_optic_Arco_2D {
+        displayName = CSTRING(optic_arco_pip);
+    };
+
+    class optic_LRPS;
+    class ACE_optic_LRPS_2D: optic_LRPS {
+        displayName = CSTRING(optic_lrps_2d);
+    };
+    class ACE_optic_LRPS_PIP: ACE_optic_LRPS_2D {
+        displayName = CSTRING(optic_lrps_pip);
+    };
+
+    class optic_MRCO;
+    class ACE_optic_MRCO_2D: optic_MRCO {
+        displayName = CSTRING(optic_MRCO_2d);
+    };
+    class ACE_optic_MRCO_PIP: ACE_optic_MRCO_2D {
+        displayName = CSTRING(optic_MRCO_pip);
+    };
+};
