+++ addons/realisticnames/compat_optics/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT optics
+#define SUBCOMPONENT_BEAUTIFIED Optics
+#include "..\script_component.hpp"
