+++ addons/realisticnames/compat_vehicles/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT vehicles
+#define SUBCOMPONENT_BEAUTIFIED Vehicles
+#include "..\script_component.hpp"
